major update of database-structure
The database now consists of three tables: reminders, tags and remindersToTags. The reminders-table now longer contains the tags-column and the connection between a reminder and its tags are now saved in the remindersToTags-table. When a reminder is created a new row is created in the remindersToTags-table for each of the new reminder's tags.